Action return VideoObservation, Renderer do not display base64, base64 in utils, makedirs before env init, ffmpeg readme
