+++ examples/gaia_agent/README.md
@@ -23,6 +23,9 @@ First, you need to install the dependencies for file converters and make folders
 ```bash
 pip install -r requirements.converters.txt
 ```
+
+Then, make sure you have `FFmpeg` installed (more details [here](https://github.com/kkroening/ffmpeg-python?tab=readme-ov-file#installing-ffmpeg)).
+
 Then you can run the agent using the following commands:
 - `python -m examples.gaia_agent.scripts.studio` - Interactive GUI that allows you to set the task for the agent and observe how it solves it step by step.
 - `python -m examples.gaia_agent.scripts.evaluate` - script to run evaluation on the GAIA validation set.
