+++ tapeagents/utils.py
@@ -1,5 +1,7 @@
+import base64
 import difflib
 import json
+import os
 from typing import Any
 
 import jsonref
@@ -93,3 +95,16 @@ def get_step_schemas_from_union_type(cls) -> str:
         step["properties"]["kind"] = {"const": step["properties"]["kind"]["const"]}
         clean_schema.append(step)
     return json.dumps(clean_schema, ensure_ascii=False)
+
+
+def image_base64_message(image_path: str) -> dict:
+    _, image_extension = os.path.splitext(image_path)
+    content_type = f"image/{image_extension}"
+    base64_image = encode_image(image_path)
+    message = {"type": "image_url", "image_url": {"url": f"data:{content_type};base64,{base64_image}"}}
+    return message
+
+
+def encode_image(image_path):
+    with open(image_path, "rb") as image_file:
+        return base64.b64encode(image_file.read()).decode("utf-8")
